chore: extract mockdata package
