+++ e2e/tsconfig.json
@@ -1,4 +1,8 @@
 {
     "extends": "../common/shared/tsconfigs/node.json",
-    "include": ["**/*.ts", "e2e.d.ts"]
+    "include": [
+        "**/*.ts",
+        "e2e.d.ts",
+        "../../../examples/**/*.ts"
+    ]
 }
