+++ examples/src/slides/main.ts
@@ -22,10 +22,10 @@ import { UniverDrawingPlugin } from '@univerjs/drawing';
 import { UniverFormulaEnginePlugin } from '@univerjs/engine-formula';
 
 import { UniverRenderEnginePlugin } from '@univerjs/engine-render';
+import { DEFAULT_SLIDE_DATA } from '@univerjs/mockdata';
 import { UniverSlidesPlugin } from '@univerjs/slides';
 import { UniverSlidesUIPlugin } from '@univerjs/slides-ui';
 import { UniverUIPlugin } from '@univerjs/ui';
-import { DEFAULT_SLIDE_DATA } from '../data';
 import { enUS, ruRU, zhCN } from '../locales';
 
 // univer
