+++ mockdata/README.md
@@ -1,0 +1,16 @@
+# @univerjs/mockdata
+
+[![npm version](https://img.shields.io/npm/v/@univerjs/mockdata)](https://npmjs.org/packages/@univerjs/mockdata)
+[![license](https://img.shields.io/npm/l/@univerjs/mockdata)](https://img.shields.io/npm/l/@univerjs/mockdata)
+
+## Introduction
+
+This package provides mock data for demo and e2e tests.
+
+## Usage
+
+### Installation
+
+```shell
+npm i @univerjs/mockdata
+```
