+++ mockdata/package.json
@@ -1,0 +1,175 @@
+{
+    "name": "@univerjs/mockdata",
+    "version": "0.0.1",
+    "private": true,
+    "description": "",
+    "author": "DreamNum <developer@univer.ai>",
+    "license": "Apache-2.0",
+    "funding": {
+        "type": "opencollective",
+        "url": "https://opencollective.com/univer"
+    },
+    "homepage": "https://univer.ai",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/dream-num/univer"
+    },
+    "bugs": {
+        "url": "https://github.com/dream-num/univer/issues"
+    },
+    "keywords": [],
+    "exports": {
+        ".": "./src/index.ts",
+        "./*": "./src/*"
+    },
+    "main": "./lib/cjs/index.js",
+    "module": "./lib/es/index.js",
+    "types": "./lib/types/index.d.ts",
+    "publishConfig": {
+        "access": "public",
+        "main": "./lib/cjs/index.js",
+        "module": "./lib/es/index.js",
+        "exports": {
+            ".": {
+                "import": "./lib/es/index.js",
+                "require": "./lib/cjs/index.js",
+                "types": "./lib/types/index.d.ts"
+            },
+            "./*": {
+                "import": "./lib/es/*",
+                "require": "./lib/cjs/*",
+                "types": "./lib/types/index.d.ts"
+            },
+            "./lib/*": "./lib/*"
+        }
+    },
+    "directories": {
+        "lib": "lib"
+    },
+    "files": [
+        "lib"
+    ],
+    "scripts": {
+        "test": "vitest run",
+        "test:watch": "vitest",
+        "coverage": "vitest run --coverage",
+        "lint:types": "tsc --noEmit",
+        "build": "tsc && vite build"
+    },
+    "peerDependencies": {
+        "@univerjs/action-recorder": "workspace:*",
+        "@univerjs/core": "workspace:*",
+        "@univerjs/data-validation": "workspace:*",
+        "@univerjs/debugger": "workspace:*",
+        "@univerjs/design": "workspace:*",
+        "@univerjs/docs": "workspace:*",
+        "@univerjs/docs-drawing": "workspace:*",
+        "@univerjs/docs-drawing-ui": "workspace:*",
+        "@univerjs/docs-hyper-link-ui": "workspace:*",
+        "@univerjs/docs-mention-ui": "workspace:*",
+        "@univerjs/docs-thread-comment-ui": "workspace:*",
+        "@univerjs/docs-ui": "workspace:*",
+        "@univerjs/drawing": "workspace:*",
+        "@univerjs/drawing-ui": "workspace:*",
+        "@univerjs/engine-formula": "workspace:*",
+        "@univerjs/engine-render": "workspace:*",
+        "@univerjs/facade": "workspace:*",
+        "@univerjs/find-replace": "workspace:*",
+        "@univerjs/icons": "^0.1.78",
+        "@univerjs/rpc": "workspace:*",
+        "@univerjs/sheets": "workspace:*",
+        "@univerjs/sheets-conditional-formatting": "workspace:*",
+        "@univerjs/sheets-conditional-formatting-ui": "workspace:*",
+        "@univerjs/sheets-crosshair-highlight": "workspace:*",
+        "@univerjs/sheets-data-validation": "workspace:*",
+        "@univerjs/sheets-drawing": "workspace:*",
+        "@univerjs/sheets-drawing-ui": "workspace:*",
+        "@univerjs/sheets-filter": "workspace:*",
+        "@univerjs/sheets-filter-ui": "workspace:*",
+        "@univerjs/sheets-find-replace": "workspace:*",
+        "@univerjs/sheets-formula": "workspace:*",
+        "@univerjs/sheets-hyper-link": "workspace:*",
+        "@univerjs/sheets-hyper-link-ui": "workspace:*",
+        "@univerjs/sheets-numfmt": "workspace:*",
+        "@univerjs/sheets-sort": "workspace:*",
+        "@univerjs/sheets-sort-ui": "workspace:*",
+        "@univerjs/sheets-thread-comment": "workspace:*",
+        "@univerjs/sheets-thread-comment-base": "workspace:*",
+        "@univerjs/sheets-ui": "workspace:*",
+        "@univerjs/sheets-zen-editor": "workspace:*",
+        "@univerjs/slides": "workspace:*",
+        "@univerjs/slides-ui": "workspace:*",
+        "@univerjs/thread-comment": "workspace:*",
+        "@univerjs/thread-comment-ui": "workspace:*",
+        "@univerjs/ui": "workspace:*",
+        "@univerjs/uni-docs-ui": "workspace:*",
+        "@univerjs/uni-formula": "workspace:*",
+        "@univerjs/uni-formula-ui": "workspace:*",
+        "@univerjs/uni-sheets-ui": "workspace:*",
+        "@univerjs/uni-slides-ui": "workspace:*",
+        "@univerjs/uniscript": "workspace:*",
+        "@univerjs/uniui": "workspace:*",
+        "rxjs": ">=7.0.0"
+    },
+    "dependencies": {
+        "@univerjs/action-recorder": "workspace:*",
+        "@univerjs/core": "workspace:*",
+        "@univerjs/data-validation": "workspace:*",
+        "@univerjs/debugger": "workspace:*",
+        "@univerjs/design": "workspace:*",
+        "@univerjs/docs": "workspace:*",
+        "@univerjs/docs-drawing": "workspace:*",
+        "@univerjs/docs-drawing-ui": "workspace:*",
+        "@univerjs/docs-hyper-link-ui": "workspace:*",
+        "@univerjs/docs-mention-ui": "workspace:*",
+        "@univerjs/docs-thread-comment-ui": "workspace:*",
+        "@univerjs/docs-ui": "workspace:*",
+        "@univerjs/drawing": "workspace:*",
+        "@univerjs/drawing-ui": "workspace:*",
+        "@univerjs/engine-formula": "workspace:*",
+        "@univerjs/engine-render": "workspace:*",
+        "@univerjs/facade": "workspace:*",
+        "@univerjs/find-replace": "workspace:*",
+        "@univerjs/icons": "^0.1.78",
+        "@univerjs/rpc": "workspace:*",
+        "@univerjs/sheets": "workspace:*",
+        "@univerjs/sheets-conditional-formatting": "workspace:*",
+        "@univerjs/sheets-conditional-formatting-ui": "workspace:*",
+        "@univerjs/sheets-crosshair-highlight": "workspace:*",
+        "@univerjs/sheets-data-validation": "workspace:*",
+        "@univerjs/sheets-drawing": "workspace:*",
+        "@univerjs/sheets-drawing-ui": "workspace:*",
+        "@univerjs/sheets-filter": "workspace:*",
+        "@univerjs/sheets-filter-ui": "workspace:*",
+        "@univerjs/sheets-find-replace": "workspace:*",
+        "@univerjs/sheets-formula": "workspace:*",
+        "@univerjs/sheets-hyper-link": "workspace:*",
+        "@univerjs/sheets-hyper-link-ui": "workspace:*",
+        "@univerjs/sheets-numfmt": "workspace:*",
+        "@univerjs/sheets-sort": "workspace:*",
+        "@univerjs/sheets-sort-ui": "workspace:*",
+        "@univerjs/sheets-thread-comment": "workspace:*",
+        "@univerjs/sheets-thread-comment-base": "workspace:*",
+        "@univerjs/sheets-ui": "workspace:*",
+        "@univerjs/sheets-zen-editor": "workspace:*",
+        "@univerjs/slides": "workspace:*",
+        "@univerjs/slides-ui": "workspace:*",
+        "@univerjs/thread-comment": "workspace:*",
+        "@univerjs/thread-comment-ui": "workspace:*",
+        "@univerjs/ui": "workspace:*",
+        "@univerjs/uni-docs-ui": "workspace:*",
+        "@univerjs/uni-formula": "workspace:*",
+        "@univerjs/uni-formula-ui": "workspace:*",
+        "@univerjs/uni-sheets-ui": "workspace:*",
+        "@univerjs/uni-slides-ui": "workspace:*",
+        "@univerjs/uniscript": "workspace:*",
+        "@univerjs/uniui": "workspace:*"
+    },
+    "devDependencies": {
+        "@univerjs-infra/shared": "workspace:*",
+        "rxjs": "^7.8.1",
+        "typescript": "^5.5.4",
+        "vite": "^5.4.1",
+        "vitest": "^2.0.5"
+    }
+}
