+++ mockdata/src/docs/default-document-data-dreamer.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import type { IDocumentData } from '@univerjs/core';
 import {
     BooleanNumber,
     ColumnSeparatorType,
@@ -26,7 +27,6 @@ import {
     WrapTextType,
 } from '@univerjs/core';
 import { ptToPixel } from '@univerjs/engine-render';
-import type { IDocumentData } from '@univerjs/core';
 
 export const DEFAULT_DOCUMENT_DATA_DREAMER: IDocumentData = {
     id: 'd',
