+++ mockdata/src/docs/default-document-data.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import type { IDocumentData } from '@univerjs/core';
 import {
     BaselineOffset,
     BooleanNumber,
@@ -27,7 +28,6 @@ import {
     SectionType,
     WrapTextType,
 } from '@univerjs/core';
-import type { IDocumentData } from '@univerjs/core';
 
 export const DEFAULT_DOCUMENT_DATA: IDocumentData = {
     id: 'd',
