+++ mockdata/src/sheets/demo/default-workbook-data-demo.ts
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import { BooleanNumber, DataValidationErrorStyle, DataValidationOperator, DataValidationType, LocaleType } from '@univerjs/core';
-import { DATA_VALIDATION_PLUGIN_NAME } from '@univerjs/sheets-data-validation';
-
 import type { IDataValidationRule, IDocumentData, IWorkbookData } from '@univerjs/core';
 import type { ICellHyperLink } from '@univerjs/sheets-hyper-link';
+
+import { BooleanNumber, DataValidationErrorStyle, DataValidationOperator, DataValidationType, LocaleType } from '@univerjs/core';
+import { DATA_VALIDATION_PLUGIN_NAME } from '@univerjs/sheets-data-validation';
 import { PAGE5_RICHTEXT_1 } from '../../slides/rich-text/page5-richtext1';
 
 const richTextDemo: IDocumentData = {
