+++ mockdata/src/sheets/demo/default-workbook-data-demo3.ts
@@ -14,14 +14,51 @@
  * limitations under the License.
  */
 
+import type { IDocumentData, IWorkbookData } from '@univerjs/core';
 import { BooleanNumber, BorderStyleTypes, LocaleType } from '@univerjs/core';
-import type { IWorkbookData } from '@univerjs/core';
 
-export const DEFAULT_WORKBOOK_DATA_DEMO2: IWorkbookData = {
+const richTextDemo: IDocumentData = {
+    id: 'd',
+    body: {
+        dataStream: 'Instructions: ①Project division - Fill in the specific division of labor after the project is disassembled: ②Responsible Person - Enter the responsible person\'s name here: ③Date-The specific execution time of the project (detailed to the date of a certain month), and the gray color block marks the planned real-time time of the division of labor of the project (for example, the specific execution time of [regional scene model arrangement and construction] is the 2 days marked in gray. \r\n',
+        textRuns: [
+            {
+                st: 0,
+                ed: 488,
+                ts: {
+                    cl: {
+                        rgb: 'rgb(92,92,92)',
+                    },
+                },
+            },
+        ],
+        paragraphs: [
+            {
+                startIndex: 489,
+                paragraphStyle: {
+                    spaceAbove: { v: 10 },
+                    lineSpacing: 1.2,
+                },
+            },
+        ],
+    },
+    documentStyle: {
+        pageSize: {
+            width: Number.POSITIVE_INFINITY,
+            height: Number.POSITIVE_INFINITY,
+        },
+        marginTop: 0,
+        marginBottom: 0,
+        marginRight: 2,
+        marginLeft: 2,
+    },
+};
+
+export const DEFAULT_WORKBOOK_DATA_DEMO3: IWorkbookData = {
     id: 'workbook-01',
     locale: LocaleType.ZH_CN,
     name: 'universheet',
-    sheetOrder: ['sheet-0002'],
+    sheetOrder: ['sheet-0001'],
     styles: {
         1: {
             fs: 30,
@@ -1217,732 +1254,271 @@ export const DEFAULT_WORKBOOK_DATA_DEMO2: IWorkbookData = {
     },
     appVersion: '3.0.0-alpha',
     sheets: {
-        'sheet-0002': {
-            id: 'sheet-0002',
-            name: 'sheet0003',
+        'sheet-0001': {
+            id: 'sheet-0001',
             cellData: {
                 0: {
                     0: {
-                        s: '11',
-                        v: 'Annual Work Schedule',
+                        s: '1',
+                        v: 'A Schedule of Items',
                     },
                 },
                 1: {
                     0: {
-                        s: '12',
+                        s: '2',
+                        v: 'Division of Project',
                     },
                     1: {
-                        s: '12',
+                        s: '3',
+                        v: 'Responsible Person',
                     },
-                    16: {
-                        s: '12',
+                    2: {
+                        s: '4',
+                        v: 'Date',
                     },
                 },
                 2: {
                     0: {
-                        s: '12',
+                        v: 'General Project Manager',
+                        s: '5',
                     },
                     1: {
-                        s: '13',
-                        v: '1/2',
+                        v: '@XXX',
+                        s: '6',
+                    },
+                    2: {
+                        v: 'March 1',
+                        s: '5',
+                    },
+                    3: {
+                        v: 'March 2',
+                        s: '5',
                     },
                     4: {
-                        s: '15',
-                        v: '3/8',
+                        v: 'March 3',
+                        s: '5',
+                    },
+                    5: {
+                        v: 'March 4',
+                        s: '5',
+                    },
+                    6: {
+                        v: 'March 5',
+                        s: '5',
                     },
                     7: {
-                        s: '16',
-                        v: '1/4',
+                        v: 'March 6',
+                        s: '5',
+                    },
+                    8: {
+                        v: 'March 7',
+                        s: '5',
+                    },
+                    9: {
+                        v: 'March 8',
+                        s: '5',
                     },
                     10: {
-                        s: '17',
-                        v: '1/4',
+                        v: 'March 9',
+                        s: '5',
+                    },
+                    11: {
+                        v: 'March 10',
+                        s: '5',
+                    },
+                    12: {
+                        v: 'March 11',
+                        s: '5',
                     },
                     13: {
-                        s: '18',
-                        v: '1/4',
+                        v: 'March 12',
+                        s: '5',
                     },
-                    16: {
-                        s: '12',
+                    14: {
+                        v: 'March 13',
+                        s: '5',
                     },
                 },
                 3: {
+                    0: {
+                        v: '1、Responsible Person of Model Section',
+                        s: '7',
+                    },
                     1: {
-                        s: '12',
+                        v: '@George',
+                        s: '8',
                     },
                 },
                 4: {
                     0: {
-                        s: '12',
+                        v: 'Advertisement Signboard',
+                        s: '9',
                     },
                     1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: 'Go to the party',
-                    },
-                    3: {
-                        s: '14',
-                        v: '√',
+                        v: '@Paul',
+                        s: '6',
                     },
                     4: {
-                        s: '12',
+                        s: '10',
                     },
                     5: {
-                        s: '14',
-                        v: '×××××',
+                        s: '10',
                     },
                     6: {
-                        s: '14',
-                        v: '√',
+                        s: '10',
                     },
                     7: {
-                        s: '12',
+                        s: '10',
                     },
                     8: {
-                        s: '14',
-                        v: '×××××',
+                        s: '10',
                     },
                     9: {
-                        s: '14',
-                        v: '√',
+                        s: '10',
                     },
                     10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
-                        v: '√',
+                        s: '10',
                     },
                 },
                 5: {
                     0: {
-                        s: '12',
+                        v: 'Transport Ready',
+                        s: '9',
                     },
                     1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: 'Purchase  Products',
-                    },
-                    3: {
-                        s: '14',
-                        v: '√',
-                    },
-                    4: {
-                        s: '12',
-                    },
-                    5: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    6: {
-                        s: '14',
-                        v: '√',
-                    },
-                    7: {
-                        s: '12',
-                    },
-                    8: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    9: {
-                        s: '14',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
-                        v: '√',
+                        v: '@George',
+                        s: '6',
                     },
                 },
                 6: {
                     0: {
-                        s: '12',
+                        v: '2、Head of Special Effects Section',
+                        s: '7',
                     },
                     1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    3: {
-                        s: '14',
-                        v: '√',
-                    },
-                    4: {
-                        s: '12',
-                    },
-                    5: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    6: {
-                        s: '14',
-                    },
-                    7: {
-                        s: '12',
-                    },
-                    8: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    9: {
-                        s: '14',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
+                        v: '@Paul',
+                        s: '8',
                     },
                 },
                 7: {
                     0: {
-                        s: '12',
+                        v: 'Render Output Parameter Test',
+                        s: '9',
                     },
                     1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: '×××××',
+                        v: '@Paul',
+                        s: '6',
                     },
                     3: {
-                        s: '14',
+                        s: '25',
                     },
                     4: {
-                        s: '12',
+                        s: '25',
                     },
                     5: {
-                        s: '14',
-                        v: '×××××',
+                        s: '25',
                     },
                     6: {
-                        s: '14',
+                        s: '25',
                     },
                     7: {
-                        s: '12',
+                        s: '25',
                     },
                     8: {
-                        s: '14',
-                        v: '×××××',
+                        s: '25',
                     },
                     9: {
-                        s: '14',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
+                        s: '25',
                     },
                 },
                 8: {
                     0: {
-                        s: '12',
+                        v: 'Camera Moving Mirror',
+                        s: '9',
                     },
                     1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    3: {
-                        s: '14',
-                    },
-                    4: {
-                        s: '12',
-                    },
-                    5: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    6: {
-                        s: '14',
-                    },
-                    7: {
-                        s: '12',
-                    },
-                    8: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    9: {
-                        s: '14',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
+                        v: '@Paul',
+                        s: '6',
                     },
                 },
                 9: {
+                    0: {
+                        v: '3、Responsible Person of Rendering Section',
+                        s: '7',
+                    },
                     1: {
-                        s: '12',
+                        v: '@Jennifer',
+                        s: '8',
                     },
                 },
                 10: {
-                    1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '19',
-                        v: 'January',
-                    },
-                    3: {
-                        s: '12',
-                    },
-                    4: {
-                        s: '12',
-                    },
-                    5: {
-                        s: '20',
-                        v: 'February',
-                    },
-                    6: {
-                        s: '12',
+                    0: {
+                        v: 'Scene Dynamic Element Design',
+                        s: '9',
                     },
                     7: {
-                        s: '12',
+                        s: '27',
                     },
                     8: {
-                        s: '21',
-                        v: 'March',
+                        s: '27',
                     },
                     9: {
-                        s: '12',
+                        s: '27',
                     },
                     10: {
-                        s: '12',
+                        s: '27',
                     },
                     11: {
-                        s: '22',
-                        v: 'April',
-                    },
-                    12: {
-                        s: '12',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '23',
-                        v: 'May',
-                    },
-                    15: {
-                        s: '12',
+                        s: '27',
                     },
                 },
                 11: {
-                    1: {
-                        s: '12',
+                    0: {
+                        v: 'Sky Map Selection',
+                        s: '9',
                     },
                 },
                 12: {
                     0: {
-                        s: '12',
-                    },
-                    1: {
-                        s: '13',
-                        v: '1/3',
-                    },
-                    4: {
-                        s: '15',
-                        v: '3/5',
-                    },
-                    7: {
-                        s: '16',
-                        v: '1/2',
-                    },
-                    10: {
-                        s: '17',
-                        v: '3/4',
-                    },
-                    13: {
-                        s: '18',
-                        v: '5/6',
-                    },
-                    16: {
-                        s: '12',
+                        v: 'Reference Scenario Data Collection',
+                        s: '9',
                     },
                 },
                 13: {
                     0: {
-                        s: '12',
-                    },
-                    1: {
-                        s: '12',
+                        v: 'Scene Dynamic Element Design',
+                        s: '9',
                     },
                     2: {
-                        s: '14',
-                        v: 'Go to the party',
+                        s: '29',
                     },
                     3: {
-                        s: '14',
-                        v: '√',
+                        s: '29',
                     },
                     4: {
-                        s: '12',
+                        s: '29',
                     },
                     5: {
-                        s: '14',
-                        v: '×××××',
+                        s: '29',
                     },
                     6: {
-                        s: '14',
-                        v: '√',
+                        s: '29',
                     },
                     7: {
-                        s: '12',
-                    },
-                    8: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    9: {
-                        s: '14',
-                        v: '√',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
-                        v: '√',
+                        s: '29',
                     },
                 },
                 14: {
                     0: {
-                        s: '12',
-                    },
-                    1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: 'Purchase Products',
-                    },
-                    3: {
-                        s: '14',
-                        v: '√',
-                    },
-                    4: {
-                        s: '12',
-                    },
-                    5: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    6: {
-                        s: '14',
-                        v: '√',
-                    },
-                    7: {
-                        s: '12',
-                    },
-                    8: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    9: {
-                        s: '14',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
-                        v: '√',
-                    },
-                },
-                15: {
-                    0: {
-                        s: '12',
-                    },
-                    1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    3: {
-                        s: '14',
-                        v: '√',
-                    },
-                    4: {
-                        s: '12',
-                    },
-                    5: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    6: {
-                        s: '14',
-                    },
-                    7: {
-                        s: '12',
-                    },
-                    8: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    9: {
-                        s: '14',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
-                    },
-                },
-                16: {
-                    0: {
-                        s: '12',
-                    },
-                    1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    3: {
-                        s: '14',
-                    },
-                    4: {
-                        s: '12',
-                    },
-                    5: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    6: {
-                        s: '14',
-                    },
-                    7: {
-                        s: '12',
-                    },
-                    8: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    9: {
-                        s: '14',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
-                    },
-                },
-                17: {
-                    0: {
-                        s: '12',
-                    },
-                    1: {
-                        s: '12',
-                    },
-                    2: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    3: {
-                        s: '14',
-                    },
-                    4: {
-                        s: '12',
-                    },
-                    5: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    6: {
-                        s: '14',
-                    },
-                    7: {
-                        s: '12',
-                    },
-                    8: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    9: {
-                        s: '14',
-                    },
-                    10: {
-                        s: '12',
-                    },
-                    11: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    12: {
-                        s: '14',
-                    },
-                    13: {
-                        s: '12',
-                    },
-                    14: {
-                        s: '14',
-                        v: '×××××',
-                    },
-                    15: {
-                        s: '14',
-                    },
-                },
-                18: {
-                    0: {
-                        s: '12',
-                    },
-                },
-                25: {
-                    0: {
-                        s: '123',
-                        v: '·',
+                        p: richTextDemo,
                     },
                 },
             },
-            tabColor: 'yellow',
+            name: 'sheet0001',
+            tabColor: 'red',
             hidden: BooleanNumber.FALSE,
-            rowCount: 26,
-            columnCount: 17,
+            rowCount: 15,
+            columnCount: 15,
             zoomRatio: 1,
             scrollTop: 200,
             scrollLeft: 100,
@@ -1953,182 +1529,113 @@ export const DEFAULT_WORKBOOK_DATA_DEMO2: IWorkbookData = {
                     startRow: 0,
                     endRow: 0,
                     startColumn: 0,
-                    endColumn: 16,
-                },
-                {
-                    startRow: 1,
-                    endRow: 17,
-                    startColumn: 0,
-                    endColumn: 0,
-                },
-                {
-                    startRow: 1,
-                    endRow: 17,
-                    startColumn: 16,
-                    endColumn: 16,
+                    endColumn: 14,
                 },
                 {
                     startRow: 1,
                     endRow: 1,
-                    startColumn: 1,
-                    endColumn: 15,
+                    startColumn: 2,
+                    endColumn: 14,
                 },
                 {
-                    startRow: 3,
-                    endRow: 3,
-                    startColumn: 1,
-                    endColumn: 15,
-                },
-                {
-                    startRow: 2,
-                    endRow: 2,
-                    startColumn: 1,
-                    endColumn: 3,
-                },
-                {
-                    startRow: 2,
-                    endRow: 2,
-                    startColumn: 4,
-                    endColumn: 6,
-                },
-                {
-                    startRow: 2,
-                    endRow: 2,
-                    startColumn: 7,
-                    endColumn: 9,
-                },
-                {
-                    startRow: 2,
-                    endRow: 2,
-                    startColumn: 10,
-                    endColumn: 12,
-                },
-                {
-                    startRow: 2,
-                    endRow: 2,
-                    startColumn: 13,
-                    endColumn: 15,
-                },
-                {
-                    startRow: 9,
-                    endRow: 9,
-                    startColumn: 1,
-                    endColumn: 15,
-                },
-                {
-                    startRow: 11,
-                    endRow: 11,
-                    startColumn: 1,
-                    endColumn: 15,
-                },
-                {
-                    startRow: 12,
-                    endRow: 12,
-                    startColumn: 1,
-                    endColumn: 3,
-                },
-                {
-                    startRow: 12,
-                    endRow: 12,
-                    startColumn: 4,
-                    endColumn: 6,
-                },
-                {
-                    startRow: 12,
-                    endRow: 12,
-                    startColumn: 7,
-                    endColumn: 9,
-                },
-                {
-                    startRow: 12,
-                    endRow: 12,
-                    startColumn: 10,
-                    endColumn: 12,
-                },
-                {
-                    startRow: 12,
-                    endRow: 12,
-                    startColumn: 13,
-                    endColumn: 15,
-                },
-                {
-                    startRow: 18,
-                    endRow: 22,
+                    startRow: 14,
+                    endRow: 14,
                     startColumn: 0,
-                    endColumn: 16,
+                    endColumn: 14,
                 },
             ],
             rowData: {
                 0: {
                     h: 70,
                 },
+                2: {
+                    h: 20,
+                },
                 3: {
+                    h: 20,
+                },
+                4: {
+                    h: 20,
+                },
+                5: {
+                    h: 20,
+                },
+                6: {
+                    h: 20,
+                },
+                7: {
+                    h: 20,
+                },
+                8: {
                     h: 20,
                 },
                 9: {
                     h: 20,
                 },
                 10: {
-                    h: 40,
+                    h: 20,
                 },
                 11: {
                     h: 20,
+                },
+                12: {
+                    h: 20,
+                },
+                13: {
+                    h: 20,
+                },
+                14: {
+                    h: 200,
                 },
             },
             columnData: {
                 0: {
-                    w: 50,
+                    w: 250,
                 },
                 1: {
-                    w: 20,
+                    w: 130,
                 },
                 2: {
-                    w: 150,
+                    w: 60,
                 },
                 3: {
-                    w: 30,
+                    w: 60,
                 },
                 4: {
-                    w: 20,
+                    w: 60,
                 },
                 5: {
-                    w: 150,
+                    w: 60,
                 },
                 6: {
-                    w: 30,
+                    w: 60,
                 },
                 7: {
-                    w: 20,
+                    w: 60,
                 },
                 8: {
-                    w: 150,
+                    w: 60,
                 },
                 9: {
-                    w: 30,
+                    w: 60,
                 },
                 10: {
-                    w: 20,
+                    w: 60,
                 },
                 11: {
-                    w: 150,
+                    w: 60,
                 },
                 12: {
-                    w: 30,
+                    w: 60,
                 },
                 13: {
-                    w: 20,
+                    w: 60,
                 },
                 14: {
-                    w: 150,
-                },
-                15: {
-                    w: 30,
-                },
-                16: {
-                    w: 50,
+                    w: 60,
                 },
             },
-            showGridlines: 0,
+            showGridlines: 1,
             rowHeader: {
                 width: 46,
                 hidden: BooleanNumber.FALSE,
