+++ mockdata/src/sheets/demo/default-workbook-data-demo8.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import { LocaleType } from '@univerjs/core';
 import type { IWorkbookData } from '@univerjs/core';
+import { LocaleType } from '@univerjs/core';
 
 export const DEFAULT_WORKBOOK_DATA_DEMO8: IWorkbookData = {
     id: 'workbook-01',
