+++ mockdata/src/sheets/slide-workbook-data.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import { BooleanNumber, BorderStyleTypes, LocaleType } from '@univerjs/core';
 import type { IWorkbookData } from '@univerjs/core';
+import { BooleanNumber, BorderStyleTypes, LocaleType } from '@univerjs/core';
 
 import { PAGE5_RICHTEXT_1 } from '../slides/rich-text/page5-richtext1';
