+++ mockdata/src/slides/default-pages/page4.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import { BasicShapes, PageElementType, PageType } from '@univerjs/core';
 import type { IWorksheetData } from '@univerjs/core';
+import { BasicShapes, PageElementType, PageType } from '@univerjs/core';
 
 import { SLIDE_WORKBOOK_DATA } from '../../sheets/slide-workbook-data';
