+++ mockdata/src/slides/default-slide-data.ts
