+++ mockdata/src/slides/index.ts
