+++ mockdata/vite.config.ts
@@ -1,0 +1,7 @@
+import createViteConfig from '@univerjs/shared/vite';
+import pkg from './package.json';
+
+export default ({ mode }) => createViteConfig({}, {
+    mode,
+    pkg,
+});
