+++ packages-experimental/README.md
@@ -1,3 +1,4 @@
 # Experimental Packages
 
-Experimental packages in development. These packages are not meant to be published to npm or used in production.
+Experimental packages in development. These packages are not meant to be published to npm or used in production so set
+`private` to `true` in package.json files.
